Rename `httphandler` to `kuberneteshandler`
