+++ internal/kuberneteshandler/wshijacker/conn.go
