+++ internal/kuberneteshandler/wshijacker/hijacker.go
